Move maxIssues to shared
