+++ src/applications/appeals/10182/pages/contestableIssues.js
@@ -2,11 +2,9 @@ import ContestableIssuesWidget from '../components/ContestableIssuesWidget';
 
 import { ContestableIssuesAdditionalInfo } from '../../shared/content/contestableIssues';
 
-import { maxIssues } from '../validations/issues';
-
 import { hasSomeSelected } from '../../shared/utils/issues';
 import { SELECTED } from '../../shared/constants';
-import { selectionRequired } from '../../shared/validations/issues';
+import { selectionRequired, maxIssues } from '../../shared/validations/issues';
 
 /**
  * contestable issues with add issue link (list loop)
