+++ src/applications/appeals/10182/tests/validations/issues.unit.spec.js
@@ -1,34 +1,9 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import { maxIssues, maxNameLength } from '../../validations/issues';
+import { maxNameLength } from '../../validations/issues';
 
-import { MAX_LENGTH, SELECTED } from '../../../shared/constants';
-import { getDate } from '../../../shared/utils/dates';
-
-const _ = null;
-
-describe('maxIssues', () => {
-  it('should not show an error when the array length is less than max', () => {
-    const errors = { addError: sinon.spy() };
-    maxIssues(errors, _, {});
-    expect(errors.addError.called).to.be.false;
-  });
-  it('should show not show an error when the array length is greater than max', () => {
-    const errors = { addError: sinon.spy() };
-    const validDate = getDate({ offset: { months: -2 } });
-    const template = {
-      issue: 'x',
-      decisionDate: validDate,
-      [SELECTED]: true,
-    };
-    maxIssues(errors, _, _, _, _, _, {
-      contestedIssues: new Array(MAX_LENGTH.SELECTIONS).fill(template),
-      additionalIssues: [template],
-    });
-    expect(errors.addError.called).to.be.true;
-  });
-});
+import { MAX_LENGTH } from '../../../shared/constants';
 
 describe('maxNameLength', () => {
   it('should show an error when a name is too long', () => {
