+++ src/applications/appeals/995/pages/contestableIssues.js
@@ -2,11 +2,11 @@ import ContestableIssuesWidget from '../components/ContestableIssuesWidget';
 
 import { ContestableIssuesAdditionalInfo } from '../../shared/content/contestableIssues';
 
-import { checkIssues, maxIssues } from '../validations/issues';
+import { checkIssues } from '../validations/issues';
 
 import { SELECTED } from '../../shared/constants';
 import { hasSomeSelected } from '../../shared/utils/issues';
-import { selectionRequired } from '../../shared/validations/issues';
+import { selectionRequired, maxIssues } from '../../shared/validations/issues';
 
 /**
  * contestable issues with add issue link (list loop)
