+++ src/applications/appeals/995/tests/validations/issues.unit.spec.js
@@ -1,12 +1,7 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import { getDate } from '../../../shared/utils/dates';
-import {
-  checkIssues,
-  maxIssues,
-  maxNameLength,
-} from '../../validations/issues';
+import { checkIssues, maxNameLength } from '../../validations/issues';
 
 import { MAX_LENGTH, SELECTED } from '../../../shared/constants';
 
@@ -119,27 +114,6 @@ describe('checkIssues', () => {
   it('should show an error with invalid selected additional issues date', () => {
     const errors = { addError: sinon.spy() };
     checkIssues(errors, _, _, _, _, _, getData({ aiDate: '2021-?-?' }));
-    expect(errors.addError.called).to.be.true;
-  });
-});
-
-describe('maxIssues', () => {
-  it('should not show an error when the array length is less than max', () => {
-    const errors = { addError: sinon.spy() };
-    maxIssues(errors, []);
-    expect(errors.addError.called).to.be.false;
-  });
-  it('should show not show an error when the array length is greater than max', () => {
-    const errors = { addError: sinon.spy() };
-    const validDate = getDate({ offset: { months: -2 } });
-    const template = {
-      issue: 'x',
-      decisionDate: validDate,
-      [SELECTED]: true,
-    };
-    maxIssues(errors, {
-      contestedIssues: new Array(MAX_LENGTH.SELECTIONS + 1).fill(template),
-    });
     expect(errors.addError.called).to.be.true;
   });
 });
