+++ src/applications/appeals/996/tests/validations/issues.unit.spec.js
@@ -1,31 +1,9 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import { getDate } from '../../../shared/utils/dates';
-import { maxIssues, maxNameLength } from '../../validations/issues';
+import { maxNameLength } from '../../validations/issues';
 
-import { MAX_LENGTH, SELECTED } from '../../../shared/constants';
-
-describe('maxIssues', () => {
-  it('should not show an error when the array length is less than max', () => {
-    const errors = { addError: sinon.spy() };
-    maxIssues(errors, []);
-    expect(errors.addError.called).to.be.false;
-  });
-  it('should show not show an error when the array length is greater than max', () => {
-    const errors = { addError: sinon.spy() };
-    const validDate = getDate({ offset: { months: -2 } });
-    const template = {
-      issue: 'x',
-      decisionDate: validDate,
-      [SELECTED]: true,
-    };
-    maxIssues(errors, {
-      contestedIssues: new Array(MAX_LENGTH.SELECTIONS + 1).fill(template),
-    });
-    expect(errors.addError.called).to.be.true;
-  });
-});
+import { MAX_LENGTH } from '../../../shared/constants';
 
 describe('maxNameLength', () => {
   it('should show an error when a name is too long', () => {
