+++ src/applications/appeals/996/validations/issues.js
@@ -1,14 +1,5 @@
 import { issueErrorMessages } from '../content/addIssue';
-
-import { maxSelectedErrorMessage } from '../../shared/content/contestableIssues';
 import { MAX_LENGTH } from '../../shared/constants';
-import { getSelected } from '../../shared/utils/issues';
-
-export const maxIssues = (error, data) => {
-  if (getSelected(data).length > MAX_LENGTH.SELECTIONS) {
-    error.addError(maxSelectedErrorMessage);
-  }
-};
 
 export const maxNameLength = (error, data) => {
   if (data.length > MAX_LENGTH.ISSUE_NAME) {
